starting props clean up
