+++ src/lib/js/components/molecules/Calendar/AgendaList/AgendaList.tsx
@@ -14,34 +14,32 @@ import {
 } from "../../../../helpers/displayEvent";
 import ModernStandardInner from "../../ModernStandard/ModernStandardInner";
 import { Stack } from "@mui/material";
+import { Props } from "../../../organisms/LocalistView";
 
-interface Props extends StandardProps {
+interface AgendaProps extends StandardProps {
   events: EventElement[];
   dateContext: Moment;
   hideaddcal: HideType;
   truncatedescription: string;
   hidedescription: HideType;
   hideimages: HideType;
-  // hidetime: HideType;
   setShowDialog: Function;
   setEventSelected: Function;
-  listClassArray: string[];
-  wrapperClassArray: string[];
+  wrapperclass: string;
+  listclass: string;
 }
 
-function AgendaList(props: Props) {
+function AgendaList(props: AgendaProps) {
   const { setShowDialog, setEventSelected } = useContext(EventContext);
 
   if (props.events.length === 0) {
     return <p>There are no events in this range.</p>;
   }
-  const { events, listClassArray, wrapperClassArray } = props;
-  const wrapperClassList = wrapperClassArray.join(" ");
-  const listClassList = listClassArray.join(" ");
+  const { events, listclass, wrapperclass } = props;
   return (
     <section className="rlv-modern-standard" title="Events List">
-      <div className={wrapperClassList}>
-        <Stack className={listClassList} spacing={2}>
+      <div className={wrapperclass}>
+        <Stack className={listclass} spacing={2}>
           {events.length > 0 ? (
             events.map((event) => {
               const key = getEventKey(event.event);
