+++ src/lib/js/components/molecules/Cards/Cards.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { Grid } from "@mui/material";
-import { StandardProps } from "../../../../types/types";
 import {
   getClassItem,
   getEventStartMonthDayString,
@@ -8,21 +7,20 @@ import {
   getEventKey,
 } from "../../../helpers/displayEvent";
 import ModernCompactInner from "../ModernCompact/ModernCompactInner";
+import { Props } from "../../organisms/LocalistView";
 
-const Cards = (props: StandardProps) => {
-  const { events, listClassArray, wrapperClassArray } = props;
-  const wrapperClassList = wrapperClassArray.join(" ");
-  const listClassList = listClassArray.join(" ");
+const Cards = (props: Props) => {
+  const { events, listclass, wrapperclass } = props;
   return (
     <section className="rlv-cards">
-      <div className={wrapperClassList}>
+      <div className={wrapperclass}>
         <Grid container spacing={2}>
           {events.length > 0 ? (
             events.map((event) => {
               const key = getEventKey(event.event);
               const classList = getClassItem(event.event);
               return (
-                <Grid item key={key} className={listClassList} xs={12} md={4}>
+                <Grid item key={key} className={listclass} xs={12} md={4}>
                   <ModernCompactInner
                     title={event.event.title}
                     description={event.event.description}
